Organize the API (#13049)
* Organize the API

There are two independent APIs:

* WabiSabi for coinjoining and,
* Wallet required for providing services to the wallet client to operate.

* Make `/synchronize` a synchronous method
