+++ WalletWasabi/WebClients/Wasabi/WasabiClient.cs
@@ -30,14 +30,12 @@ public class WasabiClient
 	public static object TransactionCacheLock { get; } = new();
 	public static ushort ApiVersion { get; private set; } = ushort.Parse(Helpers.Constants.BackendMajorVersion);
 
-	#region batch
-
 	/// <remarks>
 	/// Throws OperationCancelledException if <paramref name="cancel"/> is set.
 	/// </remarks>
 	public async Task<SynchronizeResponse> GetSynchronizeAsync(uint256 bestKnownBlockHash, int count, EstimateSmartFeeMode? estimateMode = null, CancellationToken cancel = default)
 	{
-		string relativeUri = $"api/v{ApiVersion}/btc/batch/synchronize?bestKnownBlockHash={bestKnownBlockHash}&maxNumberOfFilters={count}";
+		string relativeUri = $"api/v{ApiVersion}/wallet/synchronize?bestKnownBlockHash={bestKnownBlockHash}&maxNumberOfFilters={count}";
 		if (estimateMode is { })
 		{
 			relativeUri = $"{relativeUri}&estimateSmartFeeMode={estimateMode}";
@@ -56,10 +54,6 @@ public class WasabiClient
 		return ret;
 	}
 
-	#endregion batch
-
-	#region blockchain
-
 	public async Task<IEnumerable<Transaction>> GetTransactionsAsync(Network network, IEnumerable<uint256> txHashes, CancellationToken cancel)
 	{
 		var allTxs = new List<Transaction>();
@@ -77,7 +71,7 @@ public class WasabiClient
 
 			using HttpResponseMessage response = await HttpClient.SendAsync(
 				HttpMethod.Get,
-				$"api/v{ApiVersion}/btc/blockchain/transaction-hexes?&transactionIds={string.Join("&transactionIds=", chunk.Select(x => x.ToString()))}",
+				$"api/v{ApiVersion}/wallet/transaction-hexes?&transactionIds={string.Join("&transactionIds=", chunk.Select(x => x.ToString()))}",
 				cancellationToken: cancel).ConfigureAwait(false);
 
 			if (response.StatusCode != HttpStatusCode.OK)
@@ -114,7 +108,7 @@ public class WasabiClient
 	public async Task BroadcastAsync(string hex)
 	{
 		using var content = new StringContent($"'{hex}'", Encoding.UTF8, "application/json");
-		using HttpResponseMessage response = await HttpClient.SendAsync(HttpMethod.Post, $"api/v{ApiVersion}/btc/blockchain/broadcast", content).ConfigureAwait(false);
+		using HttpResponseMessage response = await HttpClient.SendAsync(HttpMethod.Post, $"api/v{ApiVersion}/wallet/broadcast", content).ConfigureAwait(false);
 
 		if (response.StatusCode != HttpStatusCode.OK)
 		{
@@ -136,7 +130,7 @@ public class WasabiClient
 	{
 		using HttpResponseMessage response = await HttpClient.SendAsync(
 			HttpMethod.Get,
-			$"api/v{ApiVersion}/btc/blockchain/mempool-hashes",
+			$"api/v{ApiVersion}/wallet/mempool-hashes",
 			cancellationToken: cancel).ConfigureAwait(false);
 
 		if (response.StatusCode != HttpStatusCode.OK)
@@ -159,7 +153,7 @@ public class WasabiClient
 	{
 		using HttpResponseMessage response = await HttpClient.SendAsync(
 			HttpMethod.Get,
-			$"api/v{ApiVersion}/btc/blockchain/mempool-hashes?compactness={compactness}",
+			$"api/v{ApiVersion}/wallet/mempool-hashes?compactness={compactness}",
 			cancellationToken: cancel).ConfigureAwait(false);
 
 		if (response.StatusCode != HttpStatusCode.OK)
@@ -173,13 +167,10 @@ public class WasabiClient
 		return strings;
 	}
 
-	#endregion blockchain
-
-	#region software
 
 	public async Task<ushort> GetBackendMajorVersionAsync(CancellationToken cancel)
 	{
-		using HttpResponseMessage response = await HttpClient.SendAsync(HttpMethod.Get, "api/software/versions", cancellationToken: cancel).ConfigureAwait(false);
+		using HttpResponseMessage response = await HttpClient.SendAsync(HttpMethod.Get, $"api/v{ApiVersion}/wallet/versions", cancellationToken: cancel).ConfigureAwait(false);
 
 		if (response.StatusCode != HttpStatusCode.OK)
 		{
@@ -208,6 +199,4 @@ public class WasabiClient
 
 		return new UpdateManager.UpdateStatus(backendCompatible);
 	}
-
-	#endregion software
 }
